Safari-611.1.10/LayoutTests/platform/wpe/imported/w3c/web-platform-tests/css/css-color-adjust/meta/meta-color-scheme-presentational-hint-expected.txt


FAIL Author origin style rule overrides meta color-scheme. assert_equals: expected (string) "light" but got (undefined) undefined
